perf(ssr): isolate bfcache identity runtime
