+++ packages/vinext/src/server/app-browser-state.ts
@@ -38,15 +38,18 @@ import {
   createSnapshotPathAndSearch,
   type ClientNavigationRenderSnapshot,
 } from "vinext/shims/navigation";
-import {
-  countConsumedPathnameSegments,
-  isInvisibleSegment,
-  normalizePathnameForRouteMatch,
-  splitPathSegments,
-} from "../routing/utils.js";
+import { normalizePathnameForRouteMatch } from "../routing/utils.js";
 import { normalizePath } from "./normalize-path.js";
-import { INITIAL_BFCACHE_ID } from "./app-bfcache-id.js";
-import { isBfcacheSegmentId, type BfcacheIdMap } from "./app-history-state.js";
+import type { BfcacheIdMap } from "./app-history-state.js";
+import { createNextBfcacheIdMap } from "./app-bfcache-identity.js";
+
+export {
+  createBfcacheSegmentStateKeyMap,
+  createInitialBfcacheIdMap,
+  createNextBfcacheIdMap,
+  preserveBfcacheIdsForMergedElements,
+  type BfcacheStateKeyMap,
+} from "./app-bfcache-identity.js";
 
 export {
   createHistoryStateWithNavigationMetadata,
@@ -131,7 +134,6 @@ export type PendingNavigationCommit = {
 export type AppNavigationPayloadOrigin = Readonly<
   { origin: "fresh" } | { origin: "visited-cache" }
 >;
-type BfcacheStateKeyMap = Readonly<Record<string, string>>;
 
 export const FRESH_APP_NAVIGATION_PAYLOAD_ORIGIN: AppNavigationPayloadOrigin = {
   origin: "fresh",
@@ -160,239 +162,6 @@ type NonDispatchPendingNavigationCommitDispositionDecision = {
 type PendingNavigationCommitDispositionDecision =
   | DispatchPendingNavigationCommitDispositionDecision
   | NonDispatchPendingNavigationCommitDispositionDecision;
-
-// Monotonic within a single browser document. Full reloads reset the counter,
-// while the browser entry's document-scoped version gate prevents old history
-// ids from being restored into the new document and colliding with fresh mints.
-let nextBfcacheId = 0;
-
-function rememberBfcacheId(value: string): void {
-  // The hydration sentinel is the raw "0" value and intentionally does not
-  // advance the counter; fresh ids start at "_b_1_".
-  const match = /^_b_(\d+)_$/.exec(value);
-  if (!match) return;
-  nextBfcacheId = Math.max(nextBfcacheId, Number(match[1]));
-}
-
-function mintBfcacheId(): string {
-  nextBfcacheId += 1;
-  return `_b_${nextBfcacheId}_`;
-}
-
-function getVisibleTreePathSegments(treePath: string): string[] {
-  // Tree paths contain raw filesystem segments (route groups, parallel @slots,
-  // and "." default segments). Only URL-visible segments consume a pathname
-  // segment when deriving the identity prefix, so filter the invisible ones
-  // using the same authority the route graph uses (isInvisibleSegment). Missing
-  // @slot/"." here over-counts consumed segments and re-mints bfcache ids for
-  // segments that actually persisted across a parallel-route navigation.
-  return splitPathSegments(treePath).filter((segment) => !isInvisibleSegment(segment));
-}
-
-function getTreePathIdentityPrefix(pathname: string, treePath: string): string {
-  const pathnameSegments = splitPathSegments(pathname);
-  // countConsumedPathnameSegments is the shared, browser-safe slice of the
-  // canonical filesystem-segment → URL-segment mapping in app-route-graph.ts.
-  const consumedPathnameSegments = countConsumedPathnameSegments(
-    getVisibleTreePathSegments(treePath),
-    pathnameSegments.length,
-  );
-
-  if (consumedPathnameSegments === 0) return "/";
-  const segments = pathnameSegments.slice(0, consumedPathnameSegments);
-  return `/${segments.join("/")}`;
-}
-
-type AppElementsMetadata = ReturnType<typeof AppElementsWire.readMetadata>;
-
-/**
- * Metadata parsed once per element map, paired with a slotId→binding index so
- * per-slot identity lookups are O(1) instead of a linear `slotBindings.find`
- * scan (which made per-commit identity derivation O(slots^2)).
- */
-type ParsedAppElementsMetadata = {
-  metadata: AppElementsMetadata;
-  slotBindingsBySlotId: ReadonlyMap<string, AppElementsSlotBinding>;
-};
-
-function readAppElementsMetadata(elements: AppElements): ParsedAppElementsMetadata | null {
-  let metadata: AppElementsMetadata;
-  try {
-    metadata = AppElementsWire.readMetadata(elements);
-  } catch {
-    // Some low-level tests pass partial element maps without metadata.
-    return null;
-  }
-  const slotBindingsBySlotId = new Map<string, AppElementsSlotBinding>();
-  for (const binding of metadata.slotBindings) {
-    slotBindingsBySlotId.set(binding.slotId, binding);
-  }
-  return { metadata, slotBindingsBySlotId };
-}
-
-function createActiveSlotIdentity(
-  id: string,
-  parsed: ParsedAppElementsMetadata | null,
-): string | null {
-  const activeSlotBinding = parsed?.slotBindingsBySlotId.get(id);
-  if (activeSlotBinding?.activeRouteId != null) {
-    return `${id}@${activeSlotBinding.activeRouteId}`;
-  }
-
-  const interception = parsed?.metadata.interception;
-  if (interception?.slotId !== id) return null;
-
-  return `${id}@${interception.targetRouteId}`;
-}
-
-/**
- * Legacy bridge for deriving a bfcache segment identity from AppElements wire
- * keys. Keep wire-key parsing contained here until Vinext has a route-manifest
- * semantic authority equivalent to Next.js CacheNode/segment-cache state.
- */
-function createBfcacheSegmentIdentity(
-  id: string,
-  options: { metadata: ParsedAppElementsMetadata | null; pathname: string },
-): string | null {
-  const parsed = AppElementsWire.parseElementKey(id);
-  if (!parsed) return null;
-
-  if (parsed.kind === "page") {
-    return `${id}@${options.pathname}`;
-  }
-
-  if (parsed.kind === "slot") {
-    const activeSlotIdentity = createActiveSlotIdentity(id, options.metadata);
-    if (activeSlotIdentity !== null) return activeSlotIdentity;
-
-    return `${id}@${getTreePathIdentityPrefix(options.pathname, parsed.treePath)}`;
-  }
-
-  if (parsed.kind === "layout" || parsed.kind === "template") {
-    return `${id}@${getTreePathIdentityPrefix(options.pathname, parsed.treePath)}`;
-  }
-
-  return null;
-}
-
-function collectBfcacheSegmentIds(
-  elements: AppElements,
-  parsed?: ParsedAppElementsMetadata | null,
-): string[] {
-  const ids = new Set(Object.keys(elements));
-  // Reuse already-parsed metadata when the caller has it; only fall back to a
-  // fresh parse when metadata was not threaded in (e.g. createInitialBfcacheIdMap).
-  const metadata = parsed === undefined ? readAppElementsMetadata(elements) : parsed;
-  for (const layoutId of metadata?.metadata.layoutIds ?? []) {
-    ids.add(layoutId);
-  }
-
-  return Array.from(ids).filter(isBfcacheSegmentId);
-}
-
-export function createInitialBfcacheIdMap(elements: AppElements): BfcacheIdMap {
-  const ids: Record<string, string> = {};
-  for (const id of collectBfcacheSegmentIds(elements)) {
-    ids[id] = INITIAL_BFCACHE_ID;
-  }
-  return ids;
-}
-
-function normalizeBfcachePathname(pathname: string): string {
-  // Use the route-match normalizer so decoded delimiters like %2F remain data
-  // inside their segment instead of becoming structural path separators.
-  const normalized = normalizePath(normalizePathnameForRouteMatch(pathname));
-  return normalized.length > 1 ? normalized.replace(/\/$/, "") : normalized;
-}
-
-export function createBfcacheSegmentStateKeyMap(options: {
-  elements: AppElements;
-  pathname: string;
-}): BfcacheStateKeyMap {
-  const metadata = readAppElementsMetadata(options.elements);
-  const normalizedPathname = normalizeBfcachePathname(options.pathname);
-  const stateKeys: Record<string, string> = {};
-  for (const id of collectBfcacheSegmentIds(options.elements, metadata)) {
-    const stateKey = createBfcacheSegmentIdentity(id, {
-      metadata,
-      pathname: normalizedPathname,
-    });
-    if (stateKey !== null) {
-      stateKeys[id] = stateKey;
-    }
-  }
-  return stateKeys;
-}
-
-export function createNextBfcacheIdMap(options: {
-  current: BfcacheIdMap;
-  currentElements: AppElements;
-  currentPathname: string;
-  elements: AppElements;
-  nextPathname: string;
-  restored?: BfcacheIdMap | null;
-  reuseCurrent?: boolean;
-}): BfcacheIdMap {
-  const current = options.reuseCurrent === false ? {} : options.current;
-  for (const value of Object.values(current)) {
-    rememberBfcacheId(value);
-  }
-  for (const value of Object.values(options.restored ?? {})) {
-    rememberBfcacheId(value);
-  }
-
-  const currentMetadata = readAppElementsMetadata(options.currentElements);
-  const nextMetadata = readAppElementsMetadata(options.elements);
-  const currentPathname = normalizeBfcachePathname(options.currentPathname);
-  const nextPathname = normalizeBfcachePathname(options.nextPathname);
-  const ids: Record<string, string> = {};
-  for (const id of collectBfcacheSegmentIds(options.elements, nextMetadata)) {
-    const currentIdentity = createBfcacheSegmentIdentity(id, {
-      metadata: currentMetadata,
-      pathname: currentPathname,
-    });
-    const nextIdentity = createBfcacheSegmentIdentity(id, {
-      metadata: nextMetadata,
-      pathname: nextPathname,
-    });
-    const currentValue = currentIdentity === nextIdentity ? current[id] : undefined;
-    // History traversals restore persisted ids first, matching segments keep
-    // their current id, and newly-created segments mint a fresh opaque id.
-    // Restored ids intentionally win over identity-matching: the target entry's
-    // ids were authoritative when that entry was created, and traversal must
-    // faithfully restore them even if the segment's identity has since changed.
-    // Callers must clear restored ids before this point when traversal redirects
-    // change the target href, because stale history ids otherwise win here.
-    const value = options.restored?.[id] ?? currentValue ?? mintBfcacheId();
-    ids[id] = value;
-    rememberBfcacheId(value);
-  }
-  return ids;
-}
-
-export function preserveBfcacheIdsForMergedElements(options: {
-  elements: AppElements;
-  next: BfcacheIdMap;
-  previous: BfcacheIdMap;
-}): BfcacheIdMap {
-  const ids: Record<string, string> = {};
-  for (const id of collectBfcacheSegmentIds(options.elements)) {
-    const nextValue = options.next[id];
-    if (nextValue !== undefined) {
-      ids[id] = nextValue;
-      continue;
-    }
-
-    const previousValue = options.previous[id];
-    if (previousValue !== undefined) {
-      ids[id] = previousValue;
-      // Keep the module-level opaque-id counter ahead of restored ids so future
-      // mints cannot reuse a value after reducer-level preservation.
-      rememberBfcacheId(previousValue);
-    }
-  }
-  return ids;
-}
 
 function createOperationRecord(options: {
   id: number;
